Feature - allow custom JSON message deserialization
